contrib: fix pkg_static not working with absolute pathes
The SRC folder being relative to top contrib directory when the build starts,
if we move to other directories we need to compensate for this change.

TOPSRC_BUILT is introduced to match the contrib dir from a contrib target
folder. It will work for relative pathes and absolute pathes (where it will be
the same a TOPSRC).

pkg_static_built should be used when patching pkg-config files during the build
phase.